Extract Express.js utils to a dedicated file
